refactor: move OrderedSeq's given, delete wrong import, rename pure literals elimination methods
