+++ src/test/scala/satify/DPLLTest/DPLLTest.scala
@@ -9,7 +9,7 @@ import satify.model.CNF.*
 import satify.model.Bool.{False, True}
 import satify.update.dpll.CNFSimplification.*
 import satify.update.dpll.PartialModelUtils.*
-import satify.model.dpll.given_Ordering_Variable
+import satify.model.dpll.OrderedSeq.given_Ordering_Variable
 import satify.model.dpll.OrderedSeq.*
 import satify.model.dpll.{Decision, DecisionTree, PartialModel}
 import satify.update.dpll.DPLL.*
