+++ src/test/scala/satify/DPLLTest/PartialModelUtilsTest.scala
@@ -4,8 +4,8 @@ import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import satify.update.dpll.PartialModelUtils.*
 import satify.model.CNF.*
-import satify.model.dpll.OrderedSeq.seq
-import satify.model.dpll.{Constraint, Decision, PartialModel, given_Ordering_Variable}
+import satify.model.dpll.OrderedSeq.{seq, given_Ordering_Variable}
+import satify.model.dpll.{Constraint, Decision, PartialModel}
 import satify.model.{CNF, Variable}
 import satify.update.dpll.DPLL.dpll
